+++ src/Middleware/DispatcherInterface.php
@@ -1,0 +1,35 @@
+<?php
+
+namespace Rougin\Slytherin\Middleware;
+
+use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\ServerRequestInterface;
+
+/**
+ * Dispatcher Interface
+ *
+ * An interface for handling third party middleware dispatchers.
+ *
+ * @package Slytherin
+ * @author  Rougin Royce Gutib <rougingutib@gmail.com>
+ */
+interface DispatcherInterface
+{
+    /**
+     * Processes the specified middlewares in stack.
+     *
+     * @param  \Psr\Http\Message\RequestInterface  $request
+     * @param  \Psr\Http\Message\ResponseInterface $response
+     * @param  array                               $stack
+     * @return \Psr\Http\Message\ResponseInterface|null
+     */
+    public function __invoke(ServerRequestInterface $request, ResponseInterface $response, array $stack = array());
+
+    /**
+     * Returns the listing of middlewares included.
+     *
+     * @param  array $middlewares
+     * @return array
+     */
+    public function stack(array $middlewares = array());
+}